Fix #22223: IAE in org.openstreetmap.josm.plugins.turnrestrictions.list.TurnRestrictionsListDialog$DeleteAction.deleteRelation
This also fixes a similar IAE in TurnRestrictgionEditor.


git-svn-id: https://josm.openstreetmap.de/osmsvn/applications/editors/josm/plugins@36009 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ turnrestrictions/src/org/openstreetmap/josm/plugins/turnrestrictions/editor/TurnRestrictionEditor.java
@@ -768,7 +768,7 @@ public class TurnRestrictionEditor extends JDialog implements NavigationControle
         @Override
         public void actionPerformed(ActionEvent e) {
             Relation tr = getTurnRestriction();
-            if (tr == null || tr.getDataSet() == null) return;
+            if (tr == null || tr.getDataSet() == null || tr.isDeleted()) return;
             org.openstreetmap.josm.actions.mapmode.DeleteAction.deleteRelation(
                     getLayer(),
                     tr
